Break up main and lib
