+++ midi_light_show/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "midi_light_show"
-version = "0.2.3"
+version = "0.2.4"
 authors = ["Terkwood <38859656+Terkwood@users.noreply.github.com>"]
 edition = "2018"
